refactor(T9766): centralize logger + memory types in @cleocode/contracts (Wave 1)
T9758 saga — Wave 1 of moving cross-package types out of @cleocode/core
and into @cleocode/contracts per AGENTS.md SSoT rule.

Moved (5 types):
- LoggerConfig (new packages/contracts/src/logger.ts)
- MemorySearchHit, MemoryGraphStats (packages/contracts/src/memory.ts)
- MemoryDecisionRecord (packages/contracts/src/memory.ts — renamed from
  DecisionRecord to avoid collision with session-ops DecisionRecord that
  already lives in packages/contracts/src/operations/session.ts)
- PatternRecord, LearningRecord (packages/contracts/src/memory.ts)

Core re-exports the same names (DecisionRecord remains an alias for
MemoryDecisionRecord) so existing imports from @cleocode/core stay green —
zero breaking changes for Studio, CLI, or external consumers.

Consumers updated to import the type contract directly from
@cleocode/contracts where the type is consumed without core runtime:
- packages/cleo/src/cli/logger-bootstrap.ts (LoggerConfig)
- packages/studio/src/routes/api/memory/{find,graph,learnings,patterns,decisions}/+server.ts

Quality gates: contracts + core + cleo build green; memory test regressions
are pre-existing on main (verified by stash + retest); biome clean; all
three lint gates (lint-project-root-anti-pattern --strict, lint-core-first
--check, lint-contracts-core-ssot) pass.

Tasks: T9766
Refs: Saga T9758 (release-as-product), Epic T9752
